+++ apps/docs/src/guide/contracts/cost-estimation.md
@@ -1,6 +1,6 @@
 # Estimating Contract Call Cost
 
-The `getTransactionCost` function provided by the [Account](../../api/Account/Account.md) allows you to estimate the cost of a specific contract call. The return type, `TransactionCost`, is an object containing relevant information for the estimation:
+The [`Account.getTransactionCost`](../../api/Account/Account.md#gettransactioncost) metod allows you to estimate the cost of a specific contract call. The return type, `TransactionCost`, is an object containing relevant information for the estimation:
 
 <<< @/../../../packages/account/src/providers/provider.ts#cost-estimation-1{ts:line-numbers}
 
@@ -8,10 +8,10 @@ The following example demonstrates how to get the estimated transaction cost for
 
 ## 1. Single contract call transaction:
 
-<<< @/../../docs-snippets/src/guide/contracts/cost-estimation.test.ts#cost-estimation-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/cost-estimation.ts#cost-estimation-1{ts:line-numbers}
 
 ## 2. Multiple contract calls transaction:
 
-<<< @/../../docs-snippets/src/guide/contracts/cost-estimation.test.ts#cost-estimation-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/contracts/cost-estimation.ts#cost-estimation-2{ts:line-numbers}
 
 You can use the transaction cost estimation to set the gas limit for an actual call or display the estimated cost to the user.
